+++ substrafl/remote/register/register.py
@@ -73,8 +73,11 @@ if __name__ == "__main__":
     # Create a Substra algo from the wrapped user code
     remote_instance = remote_struct.get_remote_instance()
 
+    # Register the functions to substra-tools
+    remote_instance.register_substratools_functions()
+
     # Execute the algo using substra-tools
-    tools.execute(*remote_instance.tools_functions())
+    tools.execute()
 """